+++ internal/config/config_test.go
@@ -52,8 +52,8 @@ func TestLoadFileNotFound(t *testing.T) {
 }
 
 func TestDefault(t *testing.T) {
-	if Default.Version != "1.17.0" {
-		t.Errorf("Expected default version 1.17.0, got %s", Default.Version)
+	if Default.Version != "1.18.0" {
+		t.Errorf("Expected default version 1.18.0, got %s", Default.Version)
 	}
 }